+++ lib/pages/others/creditos-screen.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_web_browser/flutter_web_browser.dart';
 
-import 'package:moblie/utils/colors.dart';
 import 'package:moblie/utils/settings.dart';
 import 'package:moblie/widgets/ui/divider.dart';
 import 'package:moblie/widgets/ui/list_cell.dart';
@@ -30,7 +29,14 @@ class CreditoScreen extends StatelessWidget {
               ListCell(
                 title: ('Toca para contactar'),
                 subtitle: ('El objetivo de esta aplicación es ayudar a aumentar el vocabulario de las peronas que estan aprendiendo Inglés, con una interfaz limpia y  sinpublicidad.'),
-                onTap: () async => await FlutterWebBrowser.openWebPage(url: authorEmail, androidToolbarColor: primaryColor,),
+                onTap: () async => await FlutterWebBrowser.openWebPage(url: authorEmail),
+              ),
+
+              CommonTitle('API desarrollada por Joshua García'),
+              ListCell(
+                title: ('Toca para contactar'),
+                subtitle: ('El objetivo de la API es alimentar $appname con datos.'),
+                onTap: () async => await FlutterWebBrowser.openWebPage(url: authorEmail),
               ),
 
               CommonTitle('Iconos Diseñados por mynamepong'),
@@ -42,79 +48,18 @@ class CreditoScreen extends StatelessWidget {
                   width: 50.0,
                 ),
                 title: ('Diseñador del icono del hombre corriendo.'),
-                subtitle: ('Toca en el icono para mas información'),
+                subtitle: ('Icon made by mynamepong from www.flaticon.com'),
                 onTap: () => FlutterWebBrowser.openWebPage(url: 'https://www.flaticon.com/authors/mynamepong'),
               ),
 
               CommonTitle('Iconos Diseñados por Freepik'),
-              SubTitle('Toca en el icono para mas información'),
+              SubTitle('Todos los iconos de colores, excepto el hombre corriendo y el icono de la aplicación, fueron diseñados y creados por Freepik'),
               ListCell(
-                leading: Image.asset(
-                  'assets/books.png',
-                  height: 50.0,
-                  width: 50.0,
-                ),
-                title: ('Diseñador del icono del libro.'),
-                subtitle: ('Toca en el icono para mas información'),
-                onTap: () => FlutterWebBrowser.openWebPage(url: 'www.flaticon.com'),
+                title: ('Toca para obtener más información.'),
+                subtitle: ('Icons made by Freepik from www.flaticon.com'),
+                onTap: () => FlutterWebBrowser.openWebPage(url: 'https://www.flaticon.com/authors/freepik')
               ),
 
-              Separator.divider(height: 0.0, indent: 74.0),
-              ListCell(
-                leading: Image.asset(
-                  'assets/heart.png',
-                  height: 50.0,
-                  width: 50.0,
-                ),
-                title: ('Diseñador del icono del corazón.'),
-                subtitle: ('Toca en el icono para mas información'),
-                
-                onTap: () => FlutterWebBrowser.openWebPage(url: 'https://www.flaticon.com/authors/freepik'),
-              ),
-              Separator.divider(height: 0.0, indent: 74.0),
-              ListCell(
-                leading: Image.asset(
-                  'assets/shield.png',
-                  height: 50.0,
-                  width: 50.0,
-                ),
-                title: ('Diseñador del icono del escudo.'),
-                subtitle: ('Toca en el icono para mas información'),
-                onTap: () => FlutterWebBrowser.openWebPage(url: 'https://www.flaticon.com/authors/freepik'),
-              ),
-              Separator.divider(height: 0.0, indent: 74.0),
-              ListCell(
-                leading: Image.asset(
-                  'assets/info.png',
-                  height: 50.0,
-                  width: 50.0,
-                ),
-                title: ('Diseñador del icono de pregunta.'),
-                subtitle: ('Toca en el icono para mas información'),
-                onTap: () => FlutterWebBrowser.openWebPage(url: 'https://www.flaticon.com/authors/freepik'),
-              ),
-              Separator.divider(height: 0.0, indent: 74.0),
-              ListCell(
-                leading: Image.asset(
-                  'assets/idea.png',
-                  height: 50.0,
-                  width: 50.0,
-                ),
-                title: ('Diseñador del icono del foco.'),
-                subtitle: ('Toca en el icono para mas información'),
-                onTap: () => FlutterWebBrowser.openWebPage(url: 'https://www.flaticon.com/authors/freepik'),
-              ),
-              Separator.divider(height: 0.0, indent: 74.0),
-              ListCell(
-                leading: Image.asset(
-                  'assets/planet.png',
-                  height: 50.0,
-                  width: 50.0,
-                ),
-                title: ('Diseñador del icono de Planeta.'),
-                subtitle: ('Toca en el icono para mas información'),
-                onTap: () => FlutterWebBrowser.openWebPage(url: 'https://www.flaticon.com/authors/freepik'),
-              ),
             ]
           )
         )