+++ lib/pages/screens/intro-screen.dart
@@ -24,35 +24,37 @@ class _IntroScreenState extends State<IntroScreen> {
     slides.add(
        Slide(
         title: "VOCABULARIO",
-        description: "El objetivo principal de $appname es aumentar tu vocabulario.",
+        description: "Aumenta tu vocabulario con palabras al azar todo el tiempo.",
         pathImage: 'assets/books.png',
         backgroundColor: Color(0xfff5a623),
       ),
     );
+
     slides.add(
        Slide(
-        title: "SIMPLE",
-        description: "¡Con un solo toque puedes guardar tantas palabras como quieras!",
-        pathImage: 'assets/heart.png',
+        title: "MODO OFFLINE",
+        description: "Si te gusta una palabra, toca el corazón junto a ella para verla sin conexión.",
+        pathImage: 'assets/shield.png',
         backgroundColor: Color(0xff203152),
       ),
     );
+
     slides.add(
        Slide(
-        title: "A TU MANERA",
-        description: "¡Puedes aprender nuevas palabras todos los días!",
+        title: "APRENDE A TU RITMO",
+        description: "Con 15 minutos al día puedes aprender nuevas palabras.",
         pathImage: 'assets/idea.png',
         backgroundColor: Color(0XFFC41E3D),
       ),
     );
-    slides.add(
-      Slide(
-        title: "ALTA CALIDAD",
-        description: "El contenido que se muestra es lo más preciso posible, sin anuncios y una interfaz limpia.",
-        pathImage: 'assets/shield.png',
-        backgroundColor: Color(0xff9932CC),
-      ),
-    ); 
+    // slides.add(
+    //   Slide(
+    //     title: "ALTA CALIDAD",
+    //     description: "El contenido que se muestra es lo más preciso posible, sin anuncios y una interfaz limpia.",
+    //     pathImage: 'assets/shield.png',
+    //     backgroundColor: Color(0XFF732289),
+    //   ),
+    // ); 
   }
 
   @override